week 4. Broke out class into seperate header file. Used inline
