+++ kriot247/kriot247.vcxproj
@@ -141,6 +141,9 @@
   <ItemGroup>
     <ClCompile Include="kriot247.cpp" />
   </ItemGroup>
+  <ItemGroup>
+    <ClInclude Include="Question.h" />
+  </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
   </ImportGroup>
